backend/lcps_ai/requirements.txt
# AI & NLP
transformers>=4.36.0
torch>=2.1.0
sentence-transformers>=2.2.2
accelerate>=0.25.0

# Speech
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0

# Storage & Caching
diskcache>=5.6.3

# APIs & Web
requests>=2.31.0

# Task Scheduling
schedule>=1.2.0

# Utilities
python-dateutil>=2.8.2
colorama>=0.4.6
